Продолжаем работу над грамматикой Промежуточные мелочи, исправления комментов

+++ src/main/java/com/metrobot/WindowConfig.java
@@ -7,13 +7,13 @@ import java.util.Map;
 
 public class WindowConfig {
     // --- Паузы ---
-    public static final long PAUSE_LONG_MS = 4000;
-    public static final long PAUSE_SHORT_MS = 1200;
-    public static final long FIVE_MINUTES_PAUSE_SECONDS = 284;
-    public static final long PAUSE_BEFORE_BOSS_MS = 14_000;
-    public static final long PAUSE_TUNNEL_MS = 18_000;
+    public static final int PAUSE_LONG_MS = 4000;
+    public static final int PAUSE_SHORT_MS = 1200;
+    public static final int FIVE_MINUTES_PAUSE_SECONDS = 284;
+    public static final int PAUSE_BEFORE_BOSS_MS = 14_000;
+    public static final int PAUSE_TUNNEL_MS = 18_000;
 
-    // --- Максимальное количество боёв ---
+    // --- Максимально допустимое количество боёв ---
     public static final byte MAX_BATTLES_ARENA = 50;
     public static final byte MAX_BATTLES_CLANWAR = 24;
     public static final byte MAX_BATTLES_RAID = 12;
